improve performance from Type System

+++ src/functions/binary_file_io.jl
@@ -12,18 +12,18 @@ Write binary file from array data
 - `existing_file::Bool=false`: optional boolean keyword to delete or not delete existing data
 ...
 """
-function write_bin(x::Array{T,1},fileName::String;existing_file::Bool=false)::Int64 where T
+function write_bin(x::Array{T,1},fileName::String;existing_file::Bool=false)::Int where T
     if existing_file
         rm(fileName)
     end
     # Open the file
     io=open(fileName,"w")
     # Cast this number to make sure we know its type
-    write(io,Int64(size(x)[1]))
+    write(io,Int(size(x)[1]))
     # Get the type as a string
     typ=repr(T)
     # Write the length of the type string
-    write(io,Int64(length(typ)))
+    write(io,Int(length(typ)))
     # Now write the type string
     for i in eachindex(typ)
         write(io,Char(typ[i]))
@@ -49,18 +49,18 @@ Write binary file from matrix data
 - `existing_file::Bool=false`: optional boolean keyword to delete or not delete existing data
 ...
 """
-function write_bin(x::Matrix{T},fileName::String;existing_file::Bool=false)::Int64 where T
+function write_bin(x::Matrix{T},fileName::String;existing_file::Bool=false)::Int where T
     if existing_file
         rm(fileName)
     end
     # Open the file
     io=open(fileName,"w")
     # Cast this number to make sure we know its type
-    write(io,Int64(size(x)[1]))
+    write(io,Int(size(x)[1]))
     # Get the type as a string
     typ=repr(T)
     # Write the length of the type string
-    write(io,Int64(length(typ)))
+    write(io,Int(length(typ)))
     # Now write the type string
     for i in eachindex(typ)
         write(io,Char(typ[i]))
@@ -83,12 +83,12 @@ Speeds up the read binary file
 # Arguments
 - `io::IO`: in/output variable
 - `::Type{T}`: data type
-- `n::Int64`: total number of elements in array/matrix data
+- `n::Int`: total number of elements in array/matrix data
 - `matrix_data::Bool`: boolean keyword to specify matrix or array data
-- `c_dim::Int64`: column number of matrix data
+- `c_dim::Int`: column number of matrix data
 ...
 """
-function read_bin(io::IO,::Type{T},n::Int64,matrix_data::Bool,c_dim::Int64) where T
+function read_bin(io::IO,::Type{T},n::Int,matrix_data::Bool,c_dim::Int) where T
     (matrix_data==true) ? (x=Array{T,2}(undef,n,c_dim)) : (x=Array{T,1}(undef, n))
     for i in eachindex(x)
         x[i]=read(io,T)
@@ -106,16 +106,16 @@ Read binary file
 # Arguments
 - `fileName::String`: name of file data
 - `matrix_data::Bool`: optional boolean keyword to specify matrix or array data
-- `c_dim::Int64`: optional column number of matrix data
+- `c_dim::Int`: optional column number of matrix data
 ...
 """
-function read_bin(fileName::String;matrix_data=false,c_dim::Int64=1)
+function read_bin(fileName::String;matrix_data=false,c_dim::Int=1)
     # Open the file
     io=open(fileName,"r")
     # Read the total number of elements in the resulting array
-    n=read(io,Int64)
+    n=read(io,Int)
     # Read the length of the type name
-    nt=read(io,Int64)
+    nt=read(io,Int)
     # Then read the type name
     cName=Array{Char}(undef,nt)
     for i in eachindex(cName)
