+++ src/functions/default_running_input_data.jl
@@ -2,7 +2,7 @@
     `full_path_name::String`:           Full path name where you want to write problem results
     `L::Float64`:                       Finite element domain length [au]
     ´dom_type::String´                  Domain type (symetric or non-symetric domain)
-    `nev::Int64`:                       Number of eigenvalues
+    `nev::Int`:                       Number of eigenvalues
     `dimension::String`:                Dimension of eigen value problem
     `sigma::Float64`:                   Level shift used in inverse iteration [au]
     `adhoc_file_name::String`:          Julia file name with ad hoc potential
@@ -11,18 +11,18 @@
     only if dimension=="1D"
         `Δx::Float64`:                  Finite element size [au]
     only if dimension=="2D"
-        `nx::Int64`:                    Number of finite element of x direction
-        `ny::Int64`:                    Number of finite element of y direction
+        `nx::Int`:                    Number of finite element of x direction
+        `ny::Int`:                    Number of finite element of y direction
 =#
 
 function input_data(data_file_name::String)
     attributes = readdlm("$(data_file_name).dat",String);
-    init_row::Int64 = 1; # without count white spaces
-    init_column::Int64 = 3;
+    init_row::Int = 1; # without count white spaces
+    init_column::Int = 3;
     full_path_name::String          = "$(attributes[init_row,init_column])"
     L::Float64                      = parse(Float64,attributes[init_row+1,init_column])
     dom_type::String                = "$(attributes[init_row+2,init_column])"
-    nev::Int64                      = parse(Float64,attributes[init_row+3,init_column])
+    nev::Int                      = parse(Float64,attributes[init_row+3,init_column])
     dimension::String               = "$(attributes[init_row+4,init_column])"
     sigma::Float64                  = parse(Float64,attributes[init_row+5,init_column])
     adhoc_file_name::String         = "$(attributes[init_row+6,init_column])"
@@ -41,8 +41,8 @@ function input_data(data_file_name::String)
         params = Params1D("1D",L,dom_type,Δx,nev,sigma,potential_function_name,params_potential)
         different_masses = false
     elseif dimension == "2D"
-        nx = parse(Int64,attributes[init_row+12,init_column])
-        ny = parse(Int64,attributes[init_row+13,init_column])
+        nx = parse(Int,attributes[init_row+12,init_column])
+        ny = parse(Int,attributes[init_row+13,init_column])
         params = Params2D("2D",L,dom_type,nx,ny,nev,sigma,potential_function_name,params_potential)
         if "$(attributes[init_row+14,init_column])" == "false"
             different_masses = false
@@ -63,7 +63,7 @@ function create_tuple(params_potential_type_array::Vector{String},params_potenti
 end
 
 function create_tuple(analysis_param_array::Vector{String},params_potential::Tuple)
-    data_index::Int=parse(Int64,analysis_param_array[1])
+    data_index::Int=parse(Int,analysis_param_array[1])
     T::Type=typeof(params_potential[data_index])
     analysis_param = AnalysisParam(data_index,parse(T,analysis_param_array[2]),
         parse(T,analysis_param_array[3]),parse(T,analysis_param_array[4]))
@@ -76,7 +76,7 @@ function define_type(type_data::String)
     elseif type_data == "c"
         T = ComplexF64
     elseif type_data == "i"
-        T = Int64
+        T = Int
     end
     return T::Type
 end
